Add files via upload
This project involves using an ESP-32 and an SSD1306 OLED Screen to create a local weather-station display that displays an accurate time-stamp,  as well as current weather conditions/temperature in your locality. The primary focus of this project was on learning and developing FreeRTOS skills such as task creation, priority, and synchronization.