Added Project 1: Weather Station Display
This project involves using an ESP-32 and the SSD1306 OLED Screen to display weather and time/date information to the user, based on their locality. The weather data was retrieved using OpenWeatherMap to display temperature and weather condition, and the time-stamp was retrieved via NTP. The focus of this project was on developing FreeRTOS skills such as task creation, priority and synchronization as well as understanding the utilities of the ESP-32 platform.